pyproject.toml
[tool.poetry]
name = "postgres-homeworks"
version = "0.1.0"
description = "PostgerSQL homeworks"
authors = ["Dmitriy Frolov <dmitriy.a.frolov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
